[refactor] (PLENTY-3) Move responsibilities to dedicated services.
