+++ src/Services/BasketService.php
@@ -1,0 +1,56 @@
+<?php
+/**
+ * Description
+ *
+ * @license Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
+ * @copyright Copyright © 2016-present heidelpay GmbH. All rights reserved.
+ *
+ * @link  http://dev.heidelpay.com/
+ *
+ * @author  Simon Gabriel <development@heidelpay.de>
+ *
+ * @package  heidelpay/${Package}
+ */
+namespace Heidelpay\Services;
+
+use Plenty\Modules\Basket\Models\Basket;
+
+class BasketService
+{
+    /**
+     * @var LibService
+     */
+    private $libService;
+
+    /**
+     * BasketService constructor.
+     * @param LibService $libraryService
+     */
+    public function __construct(
+        LibService $libraryService
+    ) {
+        $this->libService = $libraryService;
+    }
+
+    /**
+     * Submits the Basket to the Basket-API and returns its ID.
+     *
+     * @param Basket $basket
+     * @param array  $authData
+     *
+     * @return string
+     */
+    public function requestBasketId(Basket $basket, array $authData): string
+    {
+        $params = [];
+        $params['auth'] = [
+            'login' => $authData['USER_LOGIN'],
+            'password' => $authData['USER_PWD'],
+            'senderId' => $authData['SECURITY_SENDER'],
+        ];
+        $params['basket'] = $basket->toArray();
+
+        $response = $this->libService->submitBasket($params);
+        return $response['basketId'];
+    }
+}
